Fix compilation of files that included access/tuptoaster.h
Commit 8b94dab split the old src/backend/access/heap/tuptoaster.c into three
new files src/backend/access/heap/heaptoast.c,
src/backend/access/common/detoast.c and
src/backend/access/common/toast_internals.c, but the old file had a lot of
GPDB-specific code that we need to move into these three new files.
