+++ src/include/access/toast_internals.h
@@ -41,7 +41,8 @@ extern Oid	toast_get_valid_index(Oid toastoid, LOCKMODE lock);
 
 extern void toast_delete_datum(Relation rel, Datum value, bool is_speculative);
 extern Datum toast_save_datum(Relation rel, Datum value,
-							  struct varlena *oldexternal, int options);
+							  struct varlena *oldexternal, bool isFrozen,
+							  int options);
 
 extern int	toast_open_indexes(Relation toastrel,
 							   LOCKMODE lock,
